Implement lift.GetDirection
Helper func to determine the direction of a lift, given the initial
structure of the system and lifts, this only factors in requests. If the
lift is moved to a floor by the system, to respond to a call, then this
isn't seen by the lift, until it reaches that floor and a request is
entered into the lift. This seems like a reasonable trade off for an
initial implementation, as otherwise a significant refactoring would be
required up front.